componentized contestHome page
